Port mu to Windows MSYS2 UCRT64

+++ test.zsh
@@ -354,6 +354,9 @@ quoted=$(_mu_zsh_quote_prompt $'quote " dollar $ backslash \\ newline\nnext')
 eval "roundtrip=$quoted"
 [[ "$roundtrip" == $'quote " dollar $ backslash \\ newline\nnext' ]] || fail "quoted prompt roundtrips"
 
+_mu_zsh_normalize_native_output_reply $'first\r\nsecond\r'
+[[ "$REPLY" == $'first\nsecond' ]] || fail "normalizes native CRLF output"
+
 export MU_ZSH_FAKE_LOG=${TMPDIR:-/tmp}/mu-zsh-test-${$}.log
 rm -f "$MU_ZSH_FAKE_LOG"
 MU_ZSH_OUTPUT=detail
